requirements.txt
fastapi[all]
reportlab
python-dotenv
uvicorn
pydantic
openai
pandas
pydantic-settings